kubernetes_rest_service/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Kubernetes demo RESTful service"
authors = ["stkrizh <st.krizh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
psycopg2-binary = "^2.9.1"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
ipython = "^7.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
